+++ src/api/endpoints/redirects/index.ts
@@ -1,2 +1,6 @@
 export * from "@/api/endpoints/redirects/get-redirect-logo/v1";
+export * from "@/api/endpoints/redirects/get-redirect-logo/v2";
 export * from "@/api/endpoints/redirects/get-redirect-token/v1";
+export * from "@/api/endpoints/redirects/get-redirect-token/v2";
+export * from "@/api/endpoints/redirects/get-redirect-token-image/v1";
+export * from "@/api/endpoints/redirects/get-redirect-collection-image/v1";