+++ src/jobs/collection-updates/metadata-queue.ts
@@ -1,0 +1,43 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+import { Job, Queue, QueueScheduler, Worker } from "bullmq";
+import { randomUUID } from "crypto";
+
+import { logger } from "@/common/logger";
+import { redis } from "@/common/redis";
+import { config } from "@/config/index";
+import { Collections } from "@/models/collections";
+
+const QUEUE_NAME = "collections-metadata-queue";
+
+export const queue = new Queue(QUEUE_NAME, {
+  connection: redis.duplicate(),
+  defaultJobOptions: {
+    attempts: 10,
+    removeOnComplete: 1000,
+    removeOnFail: 1000,
+  },
+});
+new QueueScheduler(QUEUE_NAME, { connection: redis.duplicate() });
+
+// BACKGROUND WORKER ONLY
+if (config.doBackgroundWork) {
+  const worker = new Worker(
+    QUEUE_NAME,
+    async (job: Job) => {
+      const { contract } = job.data;
+
+      logger.info(QUEUE_NAME, `Refresh collection metadata=${contract}`);
+      await Collections.updateCollectionMetadata(contract, "1");
+    },
+    { connection: redis.duplicate(), concurrency: 1 }
+  );
+
+  worker.on("error", (error) => {
+    logger.error(QUEUE_NAME, `Worker errored: ${error}`);
+  });
+}
+
+export const addToQueue = async (contract: string) => {
+  await queue.add(randomUUID(), { contract });
+};