+++ src/jobs/update-attribute/index.ts
@@ -1,3 +1,4 @@
 import "@/jobs/update-attribute/handle-new-sell-order";
 import "@/jobs/update-attribute/handle-new-buy-order";
 import "@/jobs/update-attribute/resync-attribute-cache";
+import "@/jobs/update-attribute/resync-attribute-floor-sell";